third_party/blink/web_tests/scrollbars/sub-pixel-overflow.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
#scroller {
  border-width: 0 1px;
  border-style:solid;
  margin-left: 26px;
  overflow: auto
}
</style>

<div id="scroller-container">
  <div id="scroller">
    <div id="content">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.</div>
  </div>
</div>

<script>
if (window.testRunner) {
  testRunner.setPageZoomFactor(1.1);
}
test(() => {
  let container = document.getElementById('scroller-container');
  let scroller = document.getElementById('scroller');
  assert_equals(container.clientHeight, scroller.clientHeight);
});
</script>
